Initial commit of our top level verilog files
To run the project:
	1.) Clone the repository
	2.) Open Lattice Diamond
	3.) Open new project
		- Open Drone2 stored in the cloned repo
	4.) Select "Tcl Console" on the bottom left of Lattice
	    Diamond.
	5.) Type the following into the Tcl Console (without the >)
		> source project_setup
	6.) This should have set up the project with the correct board info,
	    assigned test_* files to be simulation only, other .v files to be
	    simulation and synthesis, and built the project for synthesis.

This commit also has a basic setup for a project that will
hopefully be able to be shared across machines.

Also we can review this and base our style on how I have done
it here along with any comments people have.

The contents in each file (i.e. always@ blocks) is only there
so when synthesis is run it doesn't optimize the entired design
out.  These will be removed once the code has been merged into
master.  It also serves as a good test when people pull this code
and try to open it as a new project under Lattice Diamond.

Signed-off-by: bcreeley <bcreeley@pdx.edu>