Initial commit of our top level verilog files
This commit also has a basic setup for a project that will
hopefully be able to be shared across machines.

Also we can review this and base our style on how I have done
it here along with any comments people have.

The contents in each file (i.e. always@ blocks) is only there
so when synthesis is run it doesn't optimize the entired design
out.  These will be removed once the code has been merged into
master.  It also serves as a good test when people pull this code
and try to open it as a new project under Lattice Diamond.

Signed-off-by: bcreeley <bcreeley@pdx.edu>