Tested connecting receiver and pwm_generator back to back
TODO: Pull this down on my Windows box and make sure it compiles/synthesizes
      after all of the style/name changes.  Functionally nothing has changed
      (except maybe the test files), but I still need to make sure everything
      compiles.

This patch contains all of the code implemented to perform back to
back connections between the receiver and pwm_generator.  This allowed
for inputting pwm signals and then turning that into an 8-bit value, which
is sent to the pwm generator to generate 4 pwm signals to output to each
ESC/motor.

Some minor changes were made to all of these modules for this patch and
they included:
	- Changing the title/header of our team for each file.
	- Moving some defines from the receiver/pwm_generator blocks
	  into the common_defines.v file and removing some duplicate
	  parameters.
	- Changing any localparam(s) or parameter(s) to being ALL_UPPER_CASE.
	- Changed how modules were instantiated to match how it is being done
	  in the pwm_generator module
	- Added some more header information.
	- Removed any notion of Aux1, Aux2, or Mode signals because they can be
	  added in later.
	- Alot of this should have been a separate patch, but thats not how I did
	  it.