First step in buffering old IMU data and converting strobe to valid bit
Valid bit will stay on unless that data from the IMU isn't valid. For now this is a single clock cycle as new data is being latched into the output buffer. Eventually this will also look at the system calibration to determine if the data being received is worth keeping