Change clock for motor_mixer module to be us_clk instead of sys_clk
As I was updating the Level 1 Block Diagram for drone2.v I noticed
that the only two modules using the sys_clk were the IMU and the
motor_mixer. The IMU will have synchronization, but it makes sense
to just run the motor_mixer at the same clock rate as all the other
modules.

This is not tested, but it should not break anything. Regardless
it needs to be loaded on to the drone platform and see if we notice
any improvement (i.e. no output pauses).