Refactor MM to make it so it uses signed values
This removes the need to do any bit meddling and it removes
an entire state. The inputs get scaled and then the MM
equations happen directly after that. All of the equations were
changed to what the comment said that was previously above them.

This makes it so all local variables are signed within the MM.