Create method to compare actual outputs to expected outputs
This is done by creating a function that nearly mirrors
how the AC is running through its state machine. The big
difference is the function calculate_actual_rate uses multiplication
and division instead of shifting.

This is a work in progress and it would be much cleaner if each
verilog module file had all of its localparams moved into a
defines file so these defines could be used for test benches as well.

DON'T MERGE - NOT READY